# src/swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "v1.0.0",
    "title": "Supa-Charging",
    "description": "APIs to interact with open data & make calculations considering distance, cost, recharge speed of multiple EVstations plugs"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": ""
    }
  ],
  "paths": {
    "/destinations": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "latitude",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "longitude",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "range",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "time_for_charging_hours",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "battery_capacity",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "percentage_to_charge",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "outlet_types",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/getAll": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}